+++ CHANGELOG.md
@@ -32,6 +32,7 @@ This changelog also contains important changes in dependencies.
 - `usvg::Text::flattened` returns `&Group` and not `Option<&Group>` now.
 - `usvg::ImageHrefDataResolverFn` and `usvg::ImageHrefStringResolverFn`
   require `fontdb::Database` argument.
+- All shared nodes are stored in `Rc` and not `Rc<RefCell>` now.
 
 ### Removed
 - `usvg::Tree::postprocess()` and `usvg::PostProcessingSteps`. No longer needed.